Move the snapshot error requeue test to scheduler_afs_test.go
